// types.ts
export enum ProjectStatus {
  PENDING = 'Tertunda',
  PREPARATION = 'Persiapan',
  CONFIRMED = 'Dikonfirmasi',
  EDITING = 'Editing',
  PRINTING = 'Cetak',
  COMPLETED = 'Selesai',
  CANCELLED = 'Dibatalkan'
}

export enum PaymentStatus {
  LUNAS = 'Lunas',
  DP_TERBAYAR = 'DP Terbayar',
  BELUM_BAYAR = 'Belum Bayar'
}
// ...
export interface User {
    id: string;
    email: string;
    password: string;
    fullName: string;
    role: 'Admin' | 'Member';
}
// ...
export interface AddOn {
  id: string;
  name: string;
  price: number;
}

export interface TeamMember {
  id: string;
  name: string;
  role: string; // Fotografer, Videografer, Editor etc.
  email: string;
  phone: string;
  standardFee: number;
  rewardBalance: number;
}

export interface AssignedTeamMember {
  memberId: string;
  name: string;
  role: string;
  fee: number; // The fee for THIS project
  reward?: number; // The reward for THIS project
}

export interface Project {
  id: string;
  projectName: string;
  clientName: string;
  clientId: string;
  projectType: string;
  packageName: string;
  packageId: string;
  addOns: AddOn[];
  date: string;
  deadlineDate?: string;
  location: string;
  progress: number; // 0-100
  status: ProjectStatus;
  totalCost: number;
  amountPaid: number;
  paymentStatus: PaymentStatus;
  team: AssignedTeamMember[];
  notes?: string;
  accommodation?: string;
  driveLink?: string;
  startTime?: string;
  endTime?: string;
}

export enum TransactionType {
  INCOME = 'Pemasukan',
  EXPENSE = 'Pengeluaran'
}

export interface Transaction {
  id: string;
  date: string;
  description: string;
  amount: number;
  type: TransactionType;
  projectId?: string;
  category: string;
  method: 'Transfer Bank' | 'Tunai' | 'E-Wallet' | 'Sistem';
  pocketId?: string;
}
// ...
export interface NotificationSettings {
    newProject: boolean;
    paymentConfirmation: boolean;
    deadlineReminder: boolean;
}

export interface SecuritySettings {
    twoFactorEnabled: boolean;
}

export interface Profile {
    fullName: string;
    email: string;
    phone: string;
    companyName: string;
    website: string;
    address: string;
    bankAccount: string;
    bio: string;
    incomeCategories: string[];
    expenseCategories: string[];
    projectTypes: string[];
    eventTypes: string[];
    notificationSettings: NotificationSettings;
    securitySettings: SecuritySettings;
}